fixes according to code review
