+++ src/components/error-message/error-message.tsx
@@ -1,7 +1,7 @@
 import { useAppSelector } from '../../hooks';
 import styles from './error-message.module.css';
 
-function ErrorMessage(): JSX.Element | null {
+function ErrorMessage() {
   const error = useAppSelector((state) => state.error);
 
   return error ? <div className={styles.errorMessage}>{error}</div> : null;
