+++ src/components/header/header.tsx
@@ -1,6 +1,7 @@
 import { useLocation } from 'react-router-dom';
 import HeaderNavigation from '../header-navigation/header-navigation';
 import Logo from '../logo/logo';
+import { AppRoute } from '../../const';
 
 function Header() {
   const { pathname } = useLocation();
@@ -12,7 +13,7 @@ function Header() {
           <div className="header__left">
             <Logo block="header" size="large" />
           </div>
-          {pathname !== '/login' && <HeaderNavigation />}
+          {pathname !== String(AppRoute.Login) && <HeaderNavigation />}
         </div>
       </div>
     </header>
