+++ src/components/login-form/login-form.module.css
@@ -1,0 +1,6 @@
+.errorMessage {
+  color: tomato;
+  margin-top: -15px;
+  margin-bottom: 25px;
+  font-size: 12px;
+}
