+++ src/store/middlewares/redirect.ts
@@ -2,12 +2,13 @@ import { PayloadAction } from '@reduxjs/toolkit';
 import browserHistory from '../../browser-history';
 import { Middleware } from 'redux';
 import { reducer } from '../reducer';
+import { ACTION_TYPE_REDIRECT } from '../../const';
 
 type Reducer = ReturnType<typeof reducer>;
 
 export const redirect: Middleware<unknown, Reducer> =
   () => (next) => (action: PayloadAction<string>) => {
-    if (action.type === 'app/redirectToRoute') {
+    if (action.type === ACTION_TYPE_REDIRECT) {
       browserHistory.push(action.payload);
     }
